requirements.txt
crewai==0.201.1
streamlit==1.50.0
openai==1.109.1
unstructured==0.18.15
pyowm==3.5.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.78
langchain-groq==0.3.8
tools==1.0.6
requests==2.32.5
fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.9
python-dotenv==1.1.1
langchain-openai==0.3.34
numpy==2.3.3
scikit-learn==1.7.2
googletrans==4.0.2
google-cloud-translate==3.21.1
pandas==2.3.3
matplotlib==3.10.6
seaborn==0.13.2